Update README.md
Qoneqt is a platform designed to develop the creation of NFT projects on various blockchains. Qoneqt is a Launchpad, to help emerging projects get off the ground, by giving them visibility and the tools they need to get off to a good start. It's also a marketplace in the form of a raffle, so that small wallets can also own NFTs, and above all it's a project managed by the community, thanks to our 2 collections that will form the future council, which will be able to vote for the projects they consider the most promising. Our objectives are simple: to help blockchains develop their ecosystem, and to help creators launch their projects.